(wip) extract test of filter for ECJ bytecode into separate class
